baselines/SwinV2/requirements.txt
datasets==3.2.0
numpy==2.2.5
opencv_python==4.10.0.84
opencv_python_headless==4.11.0.86
pandas==1.5.3
Pillow==11.2.1
safetensors==0.5.3
scikit_learn==1.2.2
torch==2.6.0+cu126
tqdm==4.67.1
transformers==4.52.0.dev0
